# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TrendPoint {
  label: string;
  value: number;
}

export interface RegionRisk {
  region: string;
  score: number;
  affected: number;
}

export interface Activity {
  id: string;
  title: string;
  detail: string;
  time: string;
  type: string;
}

export interface RecallDatasetSummary {
  campaigns: number;
  estimatedUnitsAffected: number;
  highestExposureRecall: string;
  topVehicleModel: string;
  topComponent: string;
  sourceLabel: string;
}

export interface Dashboard {
  qualityHealth: number;
  activeIssues: number;
  criticalIssues: number;
  recallRisks: number;
  affectedVehicles: number;
  financialExposure: number;
  leadTimeDays: number;
  confidence: number;
  resolvedIssues: number;
  trend: TrendPoint[];
  modelRates: TrendPoint[];
  componentRates: TrendPoint[];
  regions: RegionRisk[];
  activity: Activity[];
  recallDataset: RecallDatasetSummary;
}

export interface QualityIssue {
  id: string;
  title: string;
  component: string;
  severity: string;
  status: string;
  riskScore: number;
  confidence: number;
  affectedVehicles: number;
  exposure: number;
  detectedDaysAgo: number;
  signal: string;
  owner: string;
}

export interface Hypothesis {
  id: string;
  title: string;
  probability: number;
  category: string;
  evidence: string[];
}

export interface TimelineEvent {
  label: string;
  detail: string;
  date: string;
  kind: string;
}

export interface GraphNode {
  id: string;
  label: string;
  value: string;
  kind: string;
}

export interface Simulator {
  affectedVehicles: number;
  failureProbability: number;
  repairCost: number;
  partsCost: number;
  laborCost: number;
  compensation: number;
  lateExposure: number;
  earlyExposure: number;
}

export interface IssueDetail {
  issue: QualityIssue;
  hypotheses: Hypothesis[];
  timeline: TimelineEvent[];
  graph: GraphNode[];
  impactedModels: TrendPoint[];
  impactedPlants: TrendPoint[];
  simulator: Simulator;
}

export interface Vehicle {
  id: string;
  model: string;
  modelYear: number;
  plant: string;
  region: string;
  mileage: number;
  software: string;
  riskScore: number;
  riskLevel: string;
  failureProbability: number;
  primaryFactor: string;
}

export interface Factor {
  label: string;
  contribution: number;
  detail: string;
}

export interface VehicleDetail {
  vehicle: Vehicle;
  factors: Factor[];
  events: Activity[];
}

export interface Investigation {
  id: string;
  issueId: string;
  title: string;
  status: string;
  severity: string;
  investigator: string;
  updatedAt: string;
}

export interface InvestigationInput {
  issueId: string;
  /** @minLength 1 */
  investigator: string;
}

export interface AssistantQuestion {
  issueId: string;
  /** @minLength 1 */
  question: string;
}

export interface AssistantAnswer {
  answer: string;
  confidence: number;
  sources: string[];
}

export interface ReportInput {
  issueId: string;
}

export interface Report {
  id: string;
  issueId: string;
  title: string;
  generatedAt: string;
  sections: string[];
}

export interface RecallSignal {
  id: string;
  title: string;
  component: string;
  severity: string;
  status: string;
  affectedVehicles: number;
  detectedDaysAgo: number;
  confidence: number;
  exposure: number;
  signal: string;
  leadingHypothesis: string;
  whyFlagged: string[];
  riskScore: number;
}

export interface RecallDatasetCoverage {
  datasetId: string;
  manufacturer: string;
  label: string;
  sourceFile: string;
  sourceDescription: string;
  coveragePeriod: string;
  synthetic: boolean;
  recordCount: number;
  estimatedUnitsAffected: number;
}

export interface RecallRadar {
  signals: RecallSignal[];
  totalAtRisk: number;
  criticalSignals: number;
  averageLeadTime: number;
  syntheticLabel: string;
  detectionTrend: TrendPoint[];
  datasetCoverage: RecallDatasetCoverage;
}

export interface RecallRecord {
  campaignId: string;
  vehicleModels: string;
  component: string;
  errorDescription: string;
  estimatedUnitsAffected: number;
  remedyAction: string;
}

export interface RecallDatasetPage {
  items: RecallRecord[];
  page: number;
  pageSize: number;
  total: number;
  totalPages: number;
  coverage: RecallDatasetCoverage;
}

export interface SoftwareVersion {
  version: string;
  fleetShare: number;
  diagnosticRate: number;
  failureRate: number;
  warrantyRate: number;
}

export interface SoftwareIntelligence {
  baselineVersion: string;
  comparisonVersion: string;
  subsystem: string;
  regressionAlert: string;
  changeInDiagnostics: number;
  changeInFailures: number;
  changeInWarranty: number;
  confidence: number;
  affectedModels: TrendPoint[];
  versions: SoftwareVersion[];
  evidence: string[];
  recommendation: string;
}

export interface OtaSimulationInput {
  /** @minimum 1 */
  population: number;
  /** @minLength 1 */
  fromVersion: string;
  /** @minLength 1 */
  toVersion: string;
  /** @minLength 1 */
  subsystem: string;
  /**
     * @minimum 1
     * @maximum 100
     */
  deploymentPercentage: number;
}

export interface OtaSimulation {
  population: number;
  deploymentPercentage: number;
  estimatedImpact: string;
  potentialAnomalies: string[];
  affectedModels: TrendPoint[];
  riskLevel: string;
  rollbackRecommendation: string;
  modelConfidence: number;
  simulatedEstimate: boolean;
}

export type ListIssuesParams = {
status?: string;
severity?: string;
};

export type ListVehiclesParams = {
search?: string;
};

export type ListRecallsParams = {
/**
 * Search campaign IDs and recall descriptions
 */
recall?: string;
/**
 * Filter by vehicle model or model year
 */
vehicle?: string;
/**
 * Filter by component or system
 */
component?: string;
/**
 * Filter by remedy action
 */
remedy?: string;
/**
 * @minimum 1
 */
page?: number;
/**
 * @minimum 1
 * @maximum 100
 */
pageSize?: number;
};

export type GetSoftwareIntelligenceParams = {
baseline?: string;
comparison?: string;
};