# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /software-intelligence:
    get:
      operationId: getSoftwareIntelligence
      tags: [software-intelligence]
      summary: Compare software versions for model-estimated regressions
      parameters:
        - in: query
          name: baseline
          schema:
            type: string
          required: false
        - in: query
          name: comparison
          schema:
            type: string
          required: false
      responses:
        "200":
          description: Software intelligence comparison
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SoftwareIntelligence"
  /ota-simulator:
    post:
      operationId: simulateOtaUpdate
      tags: [ota-simulator]
      summary: Simulate the modeled impact of an OTA update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtaSimulationInput"
      responses:
        "200":
          description: Modeled OTA simulation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtaSimulation"
components:
  schemas:
...
    TrendPoint:
      type: object
      required: [label, value]
      properties:
        label: { type: string }
        value: { type: number }
...
    SoftwareIntelligence:
      type: object
      required: [baselineVersion, comparisonVersion, subsystem, regressionAlert, changeInDiagnostics, changeInFailures, changeInWarranty, confidence, affectedModels, versions, evidence, recommendation]
      properties:
        baselineVersion: { type: string }
        comparisonVersion: { type: string }
        subsystem: { type: string }
        regressionAlert: { type: string }
        changeInDiagnostics: { type: number }
        changeInFailures: { type: number }
        changeInWarranty: { type: number }
        confidence: { type: number }
        affectedModels:
          type: array
          items: { $ref: "#/components/schemas/TrendPoint" }
        versions:
          type: array
          items: { $ref: "#/components/schemas/SoftwareVersion" }
        evidence:
          type: array
          items: { type: string }
        recommendation: { type: string }
    SoftwareVersion:
      type: object
      required: [version, fleetShare, diagnosticRate, failureRate, warrantyRate]
      properties:
        version: { type: string }
        fleetShare: { type: number }
        diagnosticRate: { type: number }
        failureRate: { type: number }
        warrantyRate: { type: number }
    OtaSimulationInput:
      type: object
      required: [population, fromVersion, toVersion, subsystem, deploymentPercentage]
      properties:
        population: { type: number, minimum: 1 }
        fromVersion: { type: string, minLength: 1 }
        toVersion: { type: string, minLength: 1 }
        subsystem: { type: string, minLength: 1 }
        deploymentPercentage: { type: number, minimum: 1, maximum: 100 }
    OtaSimulation:
      type: object
      required: [population, deploymentPercentage, estimatedImpact, potentialAnomalies, affectedModels, riskLevel, rollbackRecommendation, modelConfidence, simulatedEstimate]
      properties:
        population: { type: number }
        deploymentPercentage: { type: number }
        estimatedImpact: { type: string }
        potentialAnomalies:
          type: array
          items: { type: string }
        affectedModels:
          type: array
          items: { $ref: "#/components/schemas/TrendPoint" }
        riskLevel: { type: string }
        rollbackRecommendation: { type: string }
        modelConfidence: { type: number }
        simulatedEstimate: { type: boolean }